Fix issues with tests not finding d.ts file (renamed Journal.d.ts to DJournal.ts)
