+++ src/common/domain/DJournal.ts
