+++ src/pages/journal/journal.ts
@@ -3,14 +3,13 @@ import { IonicPage, LoadingController, NavController, NavParams, Platform, Toast
 import { Store, select } from '@ngrx/store';
 import { Observable } from 'rxjs/Observable';
 import { Subscription } from 'rxjs/Subscription';
-
+import { ExaminerWorkSchedule } from '../../common/domain/DJournal';
 import { BasePageComponent } from '../../classes/base-page';
 import { AuthenticationProvider } from '../../providers/authentication/authentication';
 import * as journalActions from './journal.actions';
 import { StoreModel } from '../../common/store.model';
 import { getTestSlots, getError, getIsLoading } from './journal.selector';
 import { getJournalState } from './journal.reducer';
-import { ExaminerWorkSchedule } from '../../common/domain/Journal';
 import { MesError } from '../../common/mes-error.model';
 import { map } from 'rxjs/operators';
 import { merge } from 'rxjs/observable/merge';
